Create metadata form, see #344
Started to implement the new metadata form

Gets the preview of the metadata file using the values from the form

Shows confirm message when editing the XML and returning to the form. Added date picker to the form. Updated the templates.

WIP Form

Updated the templates

Made the look and feel the same as the create ingest page.

Added optional description to fields

Added IDEA files to gitignore

Saves the created metadata

Added auto-generator functionality

Shows a text editor when there's no form to show

Some UI fixes. Sends the type and version of the template when saving the new metadata file.

Pom reformat test

Fixed poms

Fixed imports