Improvements to actionable; progress on #885
Make actionable buttons become disabled after they are clicked and become enabled after their action completes;
Atempt to make actions more uniform (and easier to maintain) by making actions always call the ActionImpact callback in the end of their execution, and by droping the support for actionable buttons in favour of actionable menu items;
These changes are needed to support the "Stop" JobAction, in which the button should become disabled after being clicked.

Also progress on #1275, although it is not fixed yet because actionable is not yet being used in "Edit" pages.