Fixes #1356 and fixes transferred resources move dialog (progress on #1343)
Calling 'None' action when the user clicks to follow the job, to avoid having the main page overriding the redirect.

On transferred resources:
Hide 'MoveToRoot' button if the transferred resource is in the root folder;
Exclude parent folder from 'move to' list;
Fixes NPE (client-side) on move to root;
Improves list layout when inside a dialog.