Fix #1346 by adding a back button
Improve ActionablewidgetBuilder to be able to add a back button when needed that looks like an Actionable button but is the same as the browser's back button; this button should only be visible in sidebars.
Replace breadcrumb in preservation events page with a navigation toolbar (that includes the breadcrumb).
Improve Navigationtoolbar to support showing no buttons.
Use the new "actionable" back button in show preservation event page, show preservation agent page, show risk page and show risk incidence page.