Fixes some small problems
Use "sublevel" instead of "intellectual entity" when referring to children AIP.
Buttons to create new descriptive metadata, sublevel or representation are inline.
Use the fa-plus-circle icon (related to #1386) on all create/new buttons with icon.
Move AIP actions download buttons to be together, and change their labels to be more descriptive.
Move download representation button to the end of the actions list (like it is in AIP).
Fix loading for tables, by centering the loading indicator on the table and hiding the table components while the table data is loading.
Fix risk incidence advanced search by correctly setting the config properties.
Add advanced search functionality to the tables in the "Run on" section of the "new preservation action" page.